Start a test for main ui, add a 'test helpers' module
